Remove compile warnings/errors on SGI and HP Betarun version number: 50
git-svn-id: https://svn.nfit.au.dk/svn/beta/trunk@12880 678b6752-a8a0-49ff-a511-7c2a3241fcd4

+++ betarun/C/betaerror.c
@@ -129,7 +129,7 @@ void BetaError(BetaErr err, Object *theObj)
 	/* DELIBERATELY NO BREAK HERE */
       case LeaveBasicCompErr:
 	/* Wind back to start of frame of caller of betaenv.o code stub */
-	SP = (long*)WindBackSP((long)SP, theObj, (long)thePC);
+	SP = (long*)WindBackSP((long)SP, theObj, thePC);
       }
       StackEnd = SP;
 #endif /* NEWRUN */
